pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Ixchel García <ixcheldelsun@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.1"
uvicorn = "^0.21.1"
twilio = "^8.1.0"
pyngrok = "^6.0.0"
python-dotenv = "^1.0.0"
python-multipart = "^0.0.6"
openai = "^0.27.5"
langchain = "^0.0.151"
pexpect = "^4.8.0"
motor = "^3.1.2"
pymongo = "^4.3.3"
pydantic = "^1.10.7"
pyunormalize = "^15.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
